feat: Add GQL agent schema
