+++ changes/2873.feature.md
@@ -1,0 +1,1 @@
+Add GQL `agent` type and resolver
